requirement.txt
diff-match-patch==20121119
django-appconf==1.0.2
django-ckeditor==5.6.1
django-imagekit==4.0.2
django-import-export==1.1.0
django-js-asset==1.1.0
et-xmlfile==1.0.1
jdcal==1.4
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.8
pilkit==2.0
Pillow==5.3.0
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
